archive/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "initiative-tracker"
version = "0.1.0"
description = "Phase 1 Munich student initiatives intelligence pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "requests>=2.32.3",
  "lxml>=5.3.0",
  "sqlalchemy>=2.0.36",
  "pydantic>=2.10.0",
  "typer>=0.12.5",
  "rich>=14.0.0",
  "pyyaml>=6.0.2",
  "rapidfuzz>=3.10.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
]

[project.scripts]
initiative-tracker = "initiative_tracker.cli:main"

[tool.setuptools.packages.find]
include = ["initiative_tracker*"]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"
